pyproject.toml
[tool.poetry]
name = "welbex"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
geopy = "^2.4.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
django-debug-toolbar = "^4.3.0"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.7.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 1
include_trailing_comma = true
line_length = 120
combine_as_imports = true
combine_star = true
skip_glob = ['**/migrations/**','.git','.venv']

[tool.autoflake]
in-place = true
